rpi/requirements.txt
Flask>=2.0
opencv-python-headless
gpiozero
requests
python-dotenv
